+++ apps/web/atoms/index.ts
@@ -1,5 +1,5 @@
 import { atom, selector } from 'recoil';
-import { POLAROID_HEIGHT, POLAROID_WIDTH, POSTIT_HEIHT, POSTIT_WIDTH } from './consts';
+import { POLAROID_HEIGHT, POLAROID_WIDTH, POSTIT_HEIHT, POSTIT_WIDTH } from 'web/shared/consts';
 
 // TEMP
 function generatePostIts() {
