+++ apps/web/components/canvas/shapes/index.ts
@@ -1,0 +1,4 @@
+export { Polaroids } from './Polaroids';
+export { Polaroid } from './Polaroid';
+export { PostIt } from './PostIt';
+export { PostIts } from './PostIts';
